[GH-38] Track job with tags
For now, we use the name of the job to track the ownership of the
job. But many user want to give their own job names.

It's better to use the tag to track the source of the job. And user
could always add more tags for their own usage.

Add `nf-job-id` tag to track the nf task info.