[GH-25] Support s3fs work directory.
Allow user to set s3 work directory.  We will tell float to mount the
s3fs when the work directory is in s3.

After this change, the Nextflow client doesn't need direct access to
the work directory any more.

Remove the support of globa `float` options in the `process` scope.
Remove the default task option in `float` scope.  Please specify them
in the `process` scope.

Bump the version to 0.2.0.