[GH-50] Input or output on s3 is not available.
When the `.exitcode` file is available in s3, we mark the job completed.
But sometimes due to the size of input/output file, they are not yet
synchronized to the cloud storage.  And the client may not be able to
verify those files.

The fix is to check the job status in MMC and mark the job completed when
the job is finished in MMC.

Also add some string transformation for the tag key and value.  Because
GCP only supports tags of alphabet, numbers and `-`.  The size limit is 63.

Replace `rootVolSize` with `imageVolSize` because it's size of the container's
file system.