[GH-75] Too many `put xxx into map` log
The old implementation list all jobs in the opCenter which is
unnecessary.  We should only focus on the job submitted by the
plugin.

Remove the full poll in the queue status update.  Add the job
entry when we submit it.

Remove the code related to the full job poll and clean up the logs.